% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_artists_by}
\alias{browse_artists_by}
\title{Browse artists by related id}
\usage{
browse_artists_by(entity, mbid, includes = NULL, limit = NULL,
  offset = NULL)
}
\arguments{
\item{entity}{area, recording, release, release-group or work}

\item{mbid}{id of an object, specified by entity}

\item{includes}{no includes are allowed for this function}

\item{limit}{restrict number of records retrieved from musicbrainz database}

\item{offset}{number of records to skip}
}
\value{
a tibble of artists consisting of mbid, type, name, gender, country, area and other fields
}
\description{
Browse artists by related id
}
\examples{
# browse Norwegian artists
browse_artists_by("area", "6743d351-6f37-3049-9724-5041161fff4d")

}
