% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{browse_collection_release_groups}
\alias{browse_collection_release_groups}
\title{Browse release groups in a collection}
\usage{
browse_collection_release_groups(collection, limit = NULL, offset = NULL)
}
\arguments{
\item{collection}{Collection MBID}

\item{limit}{Number of results}

\item{offset}{Result offset}
}
\description{
Browse release groups in a collection
}
