% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{browse_collection_releases}
\alias{browse_collection_releases}
\title{Browse releases in a collection}
\usage{
browse_collection_releases(collection, limit = NULL, offset = NULL)
}
\arguments{
\item{collection}{Collection MBID}

\item{limit}{Number of results}

\item{offset}{Result offset}
}
\description{
Browse releases in a collection
}
