% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_events_by}
\alias{browse_events_by}
\title{Browse events by related id}
\usage{
browse_events_by(entity, mbid, includes = NULL, limit = NULL,
  offset = NULL)
}
\arguments{
\item{entity}{area, artist or place}

\item{mbid}{id of an object, specified by entity}

\item{includes}{tags}

\item{limit}{restrict number of records retrieved from musicbrainz database}

\item{offset}{number of records to skip}
}
\value{
a tibble of events consisting of mbid, type, name, time, dates and other fields
}
\description{
Browse events by related id
}
\examples{
# browse events attended by Sting
browse_events_by("artist", "7944ed53-2a58-4035-9b93-140a71e41c34")
}
