% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_labels_by}
\alias{browse_labels_by}
\title{Browse labels by related id}
\usage{
browse_labels_by(entity, mbid, includes = NULL, limit = NULL, offset = NULL)
}
\arguments{
\item{entity}{area or release}

\item{mbid}{id of an object, specified by entity}

\item{includes}{tags}

\item{limit}{restrict number of records retrieved from musicbrainz database}

\item{offset}{number of records to skip}
}
\value{
a tibble of labels consisting of mbid, type, name, label_code, distribution area and other fields
}
\description{
Browse labels by related id
}
\examples{
# browse labels who issued Mozart's Requiem
browse_labels_by("release", "0ee7243b-ee28-4b42-8f6d-d84842d7ca60")
}
