% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_places_by}
\alias{browse_places_by}
\title{Browse places by related id}
\usage{
browse_places_by(entity, mbid, includes = NULL, limit = NULL,
  offset = NULL)
}
\arguments{
\item{entity}{area}

\item{mbid}{id of an object, specified by entity}

\item{includes}{tags}

\item{limit}{restrict number of records retrieved from musicbrainz database}

\item{offset}{number of records to skip}
}
\value{
a tibble of places consisting of mbid, type, name, address, geographical coordinates and other fields
}
\description{
Browse places by related id
}
\examples{
# browse places in Kyiv, Ukraine
browse_places_by("area", "51c526a4-ed34-4eaf-94cf-0b96cd019d47")
}
