% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_recordings_by}
\alias{browse_recordings_by}
\title{Browse recordings by related id}
\usage{
browse_recordings_by(entity, mbid, includes = NULL, limit = NULL,
  offset = NULL)
}
\arguments{
\item{entity}{artist, release}

\item{mbid}{id of an object, specified by entity}

\item{includes}{artists}

\item{limit}{restrict number of records retrieved from musicbrainz database}

\item{offset}{number of records to skip}
}
\value{
a tibble of recordings consisting of mbid, title, length and video flag. If includes are called, list column with related data will be added at the end of the results tibble
}
\description{
Browse recordings by related id
}
\examples{
# browse recordings by Auli'i Cravalho ("Moana")
browse_recordings_by("artist", "715d6b2d-9b1f-4615-8e7c-71cf5e9bcde3")
}
