% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_release_groups_by}
\alias{browse_release_groups_by}
\title{Browse release groups by related id}
\usage{
browse_release_groups_by(
  entity,
  mbid,
  includes = NULL,
  limit = NULL,
  offset = NULL
)
}
\arguments{
\item{entity}{artist, release}

\item{mbid}{id of an object, specified by entity}

\item{includes}{artists, tags}

\item{limit}{restrict number of records retrieved from musicbrainz database}

\item{offset}{number of records to skip}
}
\value{
a tibble of release groups consisting of mbid, title, primary type and first release date. If includes are called, list-column(s) with related data will be added at the end of the results tibble
}
\description{
Browse release groups by related id
}
\examples{
# browse release groups by Avril Lavigne
browse_release_groups_by("artist", "0103c1cc-4a09-4a5d-a344-56ad99a77193")
}
