% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_works_by}
\alias{browse_works_by}
\title{Browse works by related id}
\usage{
browse_works_by(entity, mbid, includes = NULL, limit = NULL,
  offset = NULL)
}
\arguments{
\item{entity}{artist}

\item{mbid}{id of an object, specified by entity}

\item{includes}{tags}

\item{limit}{restrict number of records retrieved from musicbrainz database}

\item{offset}{number of records to skip}
}
\value{
a tibble of works consisting of mbid, title, type and language.
}
\description{
Browse works by related id
}
\examples{
# browse works by Sade
browse_works_by("artist", "67930b3e-e00b-469f-8c74-fd69f20522ec")
}
