% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover-art.R
\name{cover_art_url}
\alias{cover_art_url}
\title{Cover Art URL Constructor}
\usage{
cover_art_url(mbid, size = NULL)
}
\arguments{
\item{mbid}{Release MBID}

\item{size}{Image size: 250, 500, 1200, or NULL for original}
}
\value{
URL string
}
\description{
Construct front cover art URL for a release.
}
