% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection.R
\name{get_collections_by_editor}
\alias{get_collections_by_editor}
\title{Get public collections by editor name}
\usage{
get_collections_by_editor(editor, limit = NULL, offset = NULL)
}
\arguments{
\item{editor}{Editor username}

\item{limit}{Number of results}

\item{offset}{Result offset}
}
\description{
Get public collections by editor name
}
