% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{lookup_artist_relations}
\alias{lookup_artist_relations}
\title{Lookup artist relationships}
\usage{
lookup_artist_relations(mbid, includes = NULL, format = "json")
}
\arguments{
\item{mbid}{Artist MBID}

\item{includes}{Relationship types to include (default: artist-rels)}

\item{format}{Output format}
}
\description{
Lookup artist relationships
}
