% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-extras.R
\name{lookup_disc_id}
\alias{lookup_disc_id}
\title{Lookup disc by disc ID}
\usage{
lookup_disc_id(discid, includes = NULL, format = "json")
}
\arguments{
\item{discid}{The disc ID to look up}

\item{includes}{Optional includes}

\item{format}{Output format}
}
\description{
Look up releases matching a disc ID (CD TOC).
}
