% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_entities_by_id}
\alias{lookup_entities_by_id}
\alias{lookup_area_by_id}
\alias{lookup_artist_by_id}
\alias{lookup_event_by_id}
\alias{lookup_instrument_by_id}
\alias{lookup_label_by_id}
\alias{lookup_place_by_id}
\alias{lookup_recording_by_id}
\alias{lookup_release_group_by_id}
\alias{lookup_release_by_id}
\alias{lookup_series_by_id}
\alias{lookup_work_by_id}
\title{Lookup musicbrainz entities by id}
\usage{
lookup_area_by_id(mbid, includes = NULL)

lookup_artist_by_id(mbid, includes = NULL)

lookup_event_by_id(mbid, includes = NULL)

lookup_instrument_by_id(mbid, includes = NULL)

lookup_label_by_id(mbid, includes = NULL)

lookup_place_by_id(mbid, includes = NULL)

lookup_recording_by_id(mbid, includes = NULL)

lookup_release_group_by_id(mbid, includes = NULL)

lookup_release_by_id(mbid, includes = NULL)

lookup_series_by_id(mbid, includes = NULL)

lookup_work_by_id(mbid, includes = NULL)
}
\arguments{
\item{mbid}{musicbrainz database id}

\item{includes}{character vector of names of related entities to include. See information
about avaialable includes in each of the lookup functions.}
}
\value{
Tibble containing details of the requested entity
}
\description{
Helper functions for looking up details about musicbrainz database entities
given mbid. Optional parameter `includes` allows retrieving related information.
Lookup functions are not vectorized
}
\section{Functions}{
\itemize{
\item \code{lookup_area_by_id}: lookup area by mbid.
Available includes: tags

\item \code{lookup_artist_by_id}: lookup artist by mbid.
Available includes: recordings, releases, release-groups, works and tags.

\item \code{lookup_event_by_id}: lookup event by mbid.
Available includes: tags.

\item \code{lookup_instrument_by_id}: lookup instrument by mbid.
Available includes: tags.

\item \code{lookup_label_by_id}: lookup label by mbid.
Available includes: releases and tags.

\item \code{lookup_place_by_id}: lookup artist by mbid.
Available includes: tags.

\item \code{lookup_recording_by_id}: lookup recording by mbid.
Available includes: artists, releases and tags

\item \code{lookup_release_group_by_id}: lookup release group by mbid.
Available includes: artists, releases and tags

\item \code{lookup_release_by_id}: lookup release by mbid.
Available includes: artists, labels, recordings, release-groups, and tags

\item \code{lookup_series_by_id}: lookup series by mbid.
Available includes: tags

\item \code{lookup_work_by_id}: lookup work by mbid.
Available includes: tags
}}

\examples{
# find mbid for Trondheim and lookup area information
trondheim_mbid <- search_areas("Trondheim",1)$mbid
lookup_area_by_id(trondheim_mbid)

# Lookup "The Wall" release
the_wall_mbid <- search_releases("The Wall AND artist:Pink Floyd",1)$mbid
lookup_release_by_id(the_wall_mbid)

# Lookup "David Sanborn" artist
sanborn_id <- search_artists("David+Sanborn",1)$mbid
lookup_artist_by_id(sanborn_id, includes=c("recordings"))
}
