% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{lookup_label_relations}
\alias{lookup_label_relations}
\title{Lookup label relationships}
\usage{
lookup_label_relations(mbid, includes = NULL, format = "json")
}
\arguments{
\item{mbid}{Label MBID}

\item{includes}{Relationship types to include (default: release-rels)}

\item{format}{Output format}
}
\description{
Lookup label relationships
}
