% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{lookup_recording_relations}
\alias{lookup_recording_relations}
\title{Lookup recording relationships}
\usage{
lookup_recording_relations(mbid, includes = NULL, format = "json")
}
\arguments{
\item{mbid}{Recording MBID}

\item{includes}{Relationship types to include (default: work-rels)}

\item{format}{Output format}
}
\description{
Lookup recording relationships
}
