% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-extras.R
\name{lookup_release_group_genres}
\alias{lookup_release_group_genres}
\title{Get release group genres}
\usage{
lookup_release_group_genres(mbid, format = "json")
}
\arguments{
\item{mbid}{Release group MBID}

\item{format}{Output format}
}
\description{
Get genres for a release group.
}
