% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{lookup_release_relations}
\alias{lookup_release_relations}
\title{Lookup release relationships}
\usage{
lookup_release_relations(mbid, includes = NULL, format = "json")
}
\arguments{
\item{mbid}{Release MBID}

\item{includes}{Relationship types to include (default: release-group-rels)}

\item{format}{Output format}
}
\description{
Lookup release relationships
}
