% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup-extras.R
\name{lookup_url_by_resource}
\alias{lookup_url_by_resource}
\title{Lookup URL by resource (text URL)}
\usage{
lookup_url_by_resource(resource, includes = NULL, format = "json")
}
\arguments{
\item{resource}{The URL to look up}

\item{includes}{Additional includes}

\item{format}{Output format}
}
\description{
Look up a URL entity by providing the actual URL string.
}
