% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relations.R
\name{lookup_work_relations}
\alias{lookup_work_relations}
\title{Lookup work relationships}
\usage{
lookup_work_relations(mbid, includes = NULL, format = "json")
}
\arguments{
\item{mbid}{Work MBID}

\item{includes}{Relationship types to include (default: recording-rels)}

\item{format}{Output format}
}
\description{
Lookup work relationships
}
