% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rate_limit_threshold}
\alias{rate_limit_threshold}
\title{Get or set the rate limit threshold (seconds between requests)}
\usage{
rate_limit_threshold(value)
}
\arguments{
\item{value}{New threshold in seconds. If missing, returns current value.}
}
\description{
Get or set the rate limit threshold (seconds between requests)
}
