% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\alias{search_annotations}
\alias{search_areas}
\alias{search_artists}
\alias{search_events}
\alias{search_instruments}
\alias{search_labels}
\alias{search_places}
\alias{search_recordings}
\alias{search_release_groups}
\alias{search_releases}
\alias{search_series}
\alias{search_works}
\title{Search musicbrainz database}
\usage{
search_annotations(query, limit = NULL, offset = NULL, strict = FALSE)

search_areas(query, limit = NULL, offset = NULL, strict = FALSE)

search_artists(query, limit = NULL, offset = NULL, strict = FALSE)

search_events(query, limit = NULL, offset = NULL, strict = FALSE)

search_instruments(query, limit = NULL, offset = NULL, strict = FALSE)

search_labels(query, limit = NULL, offset = NULL, strict = FALSE)

search_places(query, limit = NULL, offset = NULL, strict = FALSE)

search_recordings(query, limit = NULL, offset = NULL, strict = FALSE)

search_release_groups(query, limit = NULL, offset = NULL, strict = FALSE)

search_releases(query, limit = NULL, offset = NULL, strict = FALSE)

search_series(query, limit = NULL, offset = NULL, strict = FALSE)

search_works(query, limit = NULL, offset = NULL, strict = FALSE)
}
\arguments{
\item{query}{search string.}

\item{limit}{limit number of hits returned from database, defaults to NULL}

\item{offset}{number of hits to skip, defaults to NULL}

\item{strict}{return only exact matches with score of 100, defaults to FALSE}
}
\value{
a tibble of entities of interest
}
\description{
Perform free text search in the musicbrainz database. Scope of each of the `search` functions is limited to specific type of entity. Search string can be composed using Apache Lucene search syntax, including specifying relations between entities explicitly.
}
\section{Functions}{
\itemize{
\item \code{search_annotations}: Search annotations

\item \code{search_areas}: Search areas

\item \code{search_artists}: Search artists

\item \code{search_events}: Search events

\item \code{search_instruments}: Search instrument

\item \code{search_labels}: Search labels

\item \code{search_places}: Search places

\item \code{search_recordings}: Search recordings

\item \code{search_release_groups}: Search release groups (e.g. albums)

\item \code{search_releases}: Search releases

\item \code{search_series}: Search series

\item \code{search_works}: Search works
}}

\examples{
search_annotations("concerto")

# return only first entry
search_areas("Oslo",limit=1)

# skip first 25 entries (can be used as a follow-up query)
search_artists("George Michael", offset=25)

# return only events precisely matching given name
search_events("The Prince\\'s Trust", strict=TRUE)

}
\references{
\url{https://lucene.apache.org/core/4_3_0/queryparser/org/apache/lucene/queryparser/classic/package-summary.html#package_description}
}
