% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_events}
\alias{search_events}
\title{Search events}
\usage{
search_events(query, limit = NULL, offset = NULL, strict = FALSE)
}
\arguments{
\item{query}{search string using Lucene syntax}

\item{limit}{limit number of hits returned from database, defaults to NULL}

\item{offset}{number of hits to skip, defaults to NULL}

\item{strict}{return only exact matches with score of 100, defaults to FALSE}
}
\value{
a tibble of events
}
\description{
Search events
}
\examples{
search_events("The+Prince\\'s+Trust")

}
