% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_places}
\alias{search_places}
\title{Search places}
\usage{
search_places(query, limit = NULL, offset = NULL, strict = FALSE)
}
\arguments{
\item{query}{search string using Lucene syntax}

\item{limit}{limit number of hits returned from database, defaults to NULL}

\item{offset}{number of hits to skip, defaults to NULL}

\item{strict}{return only exact matches with score of 100, defaults to FALSE}
}
\value{
a tibble of places
}
\description{
Search places
}
\examples{
search_places("Telenor+Arena")

}
