% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_recordings}
\alias{search_recordings}
\title{Search recordings}
\usage{
search_recordings(query, limit = NULL, offset = NULL, strict = FALSE)
}
\arguments{
\item{query}{search string using Lucene syntax}

\item{limit}{limit number of hits returned from database, defaults to NULL}

\item{offset}{number of hits to skip, defaults to NULL}

\item{strict}{return only exact matches with score of 100, defaults to FALSE}
}
\value{
a tibble of recordings
}
\description{
Search recordings
}
\examples{
search_recordings("Yellow+Submarine")

}
