% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_series}
\alias{search_series}
\title{Search series}
\usage{
search_series(query, limit = NULL, offset = NULL, strict = FALSE)
}
\arguments{
\item{query}{search string using Lucene syntax}

\item{limit}{limit number of hits returned from database, defaults to NULL}

\item{offset}{number of hits to skip, defaults to NULL}

\item{strict}{return only exact matches with score of 100, defaults to FALSE}
}
\value{
a tibble of series
}
\description{
Search series
}
\examples{
search_series("The+Best")

}
