% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/musicbrainz-package.R
\docType{package}
\name{musicbrainz}
\alias{musicbrainz}
\alias{musicbrainz-package}
\title{musicbrainz}
\description{
Utilising the musicbrainz.org API, retrieve metadata about music
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ningxilai/musicbrainz,}
  \item Report bugs at \url{https://github.com/ningxilai/musicbrainz/issues}
}

}
\author{
\strong{Maintainer}: Iris Lennon \email{ningxilai@outlook.com}

Authors:
\itemize{
  \item Dmytro Perepolkin \email{dperepolkin@gmail.com}
  \item MetaBrainz Foundation (MetaBrainz)
}

}
